+++ CoreShell/CoreShellAssemblyInfo.cs
@@ -1,4 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("CoreShell.Tests")]
-[assembly: InternalsVisibleTo("CloudShell.Hosting")]
+[assembly: InternalsVisibleTo("CoreShell.Blazor")]